Reorganize some files.
