+++ lib/swagger.rb
@@ -1,9 +1,6 @@
 require 'active_record'
 require 'resque_value'
-require 'redis_impersonator'
-require 'resque_extension'
 
+require 'swagger/redis_impersonator'
+require 'swagger/resque_extension'
 require 'swagger/version'
-
-module Swagger
-end
